Update Assignment 6_Question 1.asm
Write a program that will prompt the user to enter a hex digit character (0- F), display it on the next line in decimal and ask the user if he or she wants to do it again. If the user types “Y”, the program repeats. If the user types “N” the program terminates. If the user enters an illegal character prompt the user to try again.